+++ config/webpack/webpack.config.js
@@ -51,6 +51,7 @@ const defaultWebpackConfig = (env = {}, argv = {}) => {
               ident: 'postcss',
               plugins() {
                 return [
+                  rtl(),
                   Autoprefixer({
                     browsers: [
                       'iOS >= 10',
@@ -63,7 +64,6 @@ const defaultWebpackConfig = (env = {}, argv = {}) => {
                       'last 2 safari versions',
                     ],
                   }),
-                  rtl(),
                 ];
               },
             },
@@ -132,6 +132,11 @@ const defaultWebpackConfig = (env = {}, argv = {}) => {
           cache: true,
           parallel: true,
           sourceMap: true,
+          uglifyOptions: {
+            compress: {
+              typeofs: false,
+            },
+          },
         }),
       ],
     },